feat(agents): prove runtime mutation guard capability
