+++ crates/harness-agents/src/runtime_fingerprint/probe.rs
@@ -31,7 +31,7 @@ pub(super) fn owner_run(
     }
     pidfd_self_preflight()?;
     let deadline = Instant::now() + super::RUNTIME_FINGERPRINT_PROBE_DEADLINE;
-    capability_child(deadline)?;
+    super::capability::validate(deadline)?;
     let working_directory =
         super::executable::observe_working_directory(options.working_dir(), deadline)?;
     if working_directory.fd() < 0 || working_directory.identity_digest.as_str().is_empty() {
@@ -110,232 +110,6 @@ pub(super) const CHILD_DESCRIPTOR_UNAVAILABLE: u8 = 3;
 pub(super) const CHILD_DESCRIPTOR_FAILED: u8 = 4;
 pub(super) const CHILD_GO: u8 = 5;
 
-fn capability_child(deadline: Instant) -> Result<(), RuntimeFingerprintProduceError> {
-    let role =
-        super::RuntimeOwnedChildRole::Observation(super::RuntimeObservationStage::CapabilityCheck);
-    let mut gate = [-1; 2];
-    let mut status = [-1; 2];
-    // SAFETY: both arrays contain space for exactly two descriptors.
-    if unsafe { libc::pipe2(gate.as_mut_ptr(), libc::O_CLOEXEC) } != 0 {
-        return Err(registration_error(
-            role,
-            super::RuntimeChildRegistrationStage::GateCreate,
-        ));
-    }
-    // SAFETY: both arrays contain space for exactly two descriptors.
-    if unsafe { libc::pipe2(status.as_mut_ptr(), libc::O_CLOEXEC) } != 0 {
-        close_fd(gate[0]);
-        close_fd(gate[1]);
-        return Err(registration_error(
-            role,
-            super::RuntimeChildRegistrationStage::GateCreate,
-        ));
-    }
-
-    let mut blocked = unsafe { std::mem::zeroed::<libc::sigset_t>() };
-    let mut saved = unsafe { std::mem::zeroed::<libc::sigset_t>() };
-    // SAFETY: initialized signal sets are confined to this owner thread.
-    let mask_result = unsafe {
-        libc::sigfillset(&mut blocked);
-        libc::sigdelset(&mut blocked, libc::SIGKILL);
-        libc::sigdelset(&mut blocked, libc::SIGSTOP);
-        libc::pthread_sigmask(libc::SIG_SETMASK, &blocked, &mut saved)
-    };
-    if mask_result != 0 {
-        close_pipe_pair(gate);
-        close_pipe_pair(status);
-        return Err(registration_error(
-            role,
-            super::RuntimeChildRegistrationStage::SignalIsolation,
-        ));
-    }
-
-    // SAFETY: the owner is single-threaded with respect to child ownership; the child runs only
-    // the allocation-free routine below until _exit.
-    let pid = unsafe { libc::fork() };
-    if pid == 0 {
-        child_capability_main(gate, status);
-    }
-    // SAFETY: restore the exact calling-thread mask immediately after fork.
-    let restore_result =
-        unsafe { libc::pthread_sigmask(libc::SIG_SETMASK, &saved, std::ptr::null_mut()) };
-    close_fd(gate[0]);
-    close_fd(status[1]);
-    if pid < 0 {
-        close_fd(gate[1]);
-        close_fd(status[0]);
-        return Err(registration_error(
-            role,
-            super::RuntimeChildRegistrationStage::Fork,
-        ));
-    }
-    if restore_result != 0 {
-        close_fd(gate[1]);
-        close_fd(status[0]);
-        rollback_unregistered_child(pid, deadline, role)?;
-        return Err(registration_error(
-            role,
-            super::RuntimeChildRegistrationStage::SignalIsolation,
-        ));
-    }
-
-    let child_status = match read_byte_before(status[0], deadline) {
-        Ok(status) => status,
-        Err(_) => {
-            close_fd(gate[1]);
-            close_fd(status[0]);
-            rollback_unregistered_child(
-                pid,
-                Instant::now() + super::RUNTIME_FINGERPRINT_CLEANUP_DEADLINE,
-                role,
-            )?;
-            return Err(
-                RuntimeFingerprintProduceError::ObservationDeadlineExceeded {
-                    stage: super::RuntimeObservationStage::CapabilityCheck,
-                },
-            );
-        }
-    };
-    close_fd(status[0]);
-    match child_status {
-        CHILD_READY => {}
-        CHILD_SIGNAL_FAILED => {
-            close_fd(gate[1]);
-            rollback_unregistered_child(pid, deadline, role)?;
-            return Err(registration_error(
-                role,
-                super::RuntimeChildRegistrationStage::SignalIsolation,
-            ));
-        }
-        CHILD_DESCRIPTOR_UNAVAILABLE => {
-            close_fd(gate[1]);
-            rollback_unregistered_child(pid, deadline, role)?;
-            return Err(RuntimeFingerprintProduceError::ContainmentUnavailable(
-                ContainmentUnavailableReason::DescriptorIsolationUnavailable,
-            ));
-        }
-        _ => {
-            close_fd(gate[1]);
-            rollback_unregistered_child(pid, deadline, role)?;
-            return Err(registration_error(
-                role,
-                super::RuntimeChildRegistrationStage::DescriptorIsolation,
-            ));
-        }
-    }
-
-    // SAFETY: pid is the live, still-gated direct child and flags are frozen to zero.
-    let pidfd = unsafe { libc::syscall(libc::SYS_pidfd_open, pid, 0) as libc::c_int };
-    if pidfd < 0 {
-        close_fd(gate[1]);
-        rollback_unregistered_child(pid, deadline, role)?;
-        return Err(registration_error(
-            role,
-            super::RuntimeChildRegistrationStage::PidfdOpen,
-        ));
-    }
-    if validate_ptrace_capability(pid, pidfd, gate[1], deadline).is_err() {
-        close_fd(gate[1]);
-        cleanup_registered_child(pidfd, deadline, role)?;
-        return Err(RuntimeFingerprintProduceError::ContainmentUnavailable(
-            ContainmentUnavailableReason::PostExecGuardUnavailable,
-        ));
-    }
-    close_fd(gate[1]);
-    validate_and_consume_capability_exit(pidfd, pid, deadline, role)
-}
-
-fn validate_ptrace_capability(
-    pid: libc::pid_t,
-    pidfd: libc::c_int,
-    gate: libc::c_int,
-    deadline: Instant,
-) -> Result<(), ()> {
-    let options = libc::PTRACE_O_TRACEEXEC | libc::PTRACE_O_TRACESYSGOOD;
-    // SAFETY: pid is the registered gated child and these requests do not deliver signals.
-    if unsafe { libc::ptrace(libc::PTRACE_SEIZE, pid, 0, options) } != 0
-        || unsafe { libc::ptrace(libc::PTRACE_INTERRUPT, pid, 0, 0) } != 0
-        || !waitid_pidfd_stop(pidfd, pid, deadline)
-    {
-        return Err(());
-    }
-    if write_byte(gate, CHILD_GO).is_err() {
-        return Err(());
-    }
-    // SAFETY: PTRACE_SYSCALL resumes only the registered traced child.
-    if unsafe { libc::ptrace(libc::PTRACE_SYSCALL, pid, 0, 0) } != 0
-        || !waitid_pidfd_stop(pidfd, pid, deadline)
-    {
-        return Err(());
-    }
-    let first_stop = super::syscall_guard::read_syscall_stop(pid);
-    // SAFETY: resume to the alternating tagged syscall stop.
-    if first_stop.is_none()
-        || unsafe { libc::ptrace(libc::PTRACE_SYSCALL, pid, 0, 0) } != 0
-        || !waitid_pidfd_stop(pidfd, pid, deadline)
-    {
-        return Err(());
-    }
-    let second_stop = super::syscall_guard::read_syscall_stop(pid);
-    let alternating = matches!(
-        (first_stop, second_stop),
-        (
-            Some(super::syscall_guard::SyscallStop::Entry { .. }),
-            Some(super::syscall_guard::SyscallStop::Exit)
-        ) | (
-            Some(super::syscall_guard::SyscallStop::Exit),
-            Some(super::syscall_guard::SyscallStop::Entry { .. })
-        )
-    );
-    let entry_allowed = [first_stop, second_stop]
-        .into_iter()
-        .all(|stop| match stop {
-            Some(super::syscall_guard::SyscallStop::Entry { number, arguments }) => {
-                matches!(
-                    super::syscall_guard::denied_class(number, arguments),
-                    Ok(None)
-                )
-            }
-            Some(super::syscall_guard::SyscallStop::Exit) => true,
-            None => false,
-        });
-    if !alternating
-        || !entry_allowed
-        // SAFETY: detach resumes the exact traced child without injecting a signal.
-        || unsafe { libc::ptrace(libc::PTRACE_DETACH, pid, 0, 0) } != 0
-    {
-        return Err(());
-    }
-    Ok(())
-}
-
-fn waitid_pidfd_stop(pidfd: libc::c_int, pid: libc::pid_t, deadline: Instant) -> bool {
-    loop {
-        let mut info = unsafe { std::mem::zeroed::<libc::siginfo_t>() };
-        // SAFETY: pidfd is live and the nonblocking wait writes one siginfo value.
-        if unsafe {
-            libc::waitid(
-                libc::P_PIDFD,
-                pidfd as libc::id_t,
-                &mut info,
-                libc::WSTOPPED | libc::WNOHANG,
-            )
-        } != 0
-        {
-            return false;
-        }
-        // SAFETY: waitid initialized SIGCHLD union fields when si_pid is nonzero.
-        let seen = unsafe { info.si_pid() };
-        if seen != 0 {
-            return seen == pid && info.si_code == libc::CLD_TRAPPED;
-        }
-        if Instant::now() >= deadline {
-            return false;
-        }
-        std::thread::yield_now();
-    }
-}
-
 pub(super) fn registration_error(
     role: super::RuntimeOwnedChildRole,
     stage: super::RuntimeChildRegistrationStage,
@@ -410,82 +184,6 @@ pub(super) fn cleanup_registered_child(
     })
 }
 
-fn validate_and_consume_capability_exit(
-    pidfd: libc::c_int,
-    pid: libc::pid_t,
-    deadline: Instant,
-    role: super::RuntimeOwnedChildRole,
-) -> Result<(), RuntimeFingerprintProduceError> {
-    let observed = waitid_pidfd(pidfd, true, deadline);
-    if !matches!(observed, Ok((seen, libc::CLD_EXITED, 0)) if seen == pid) {
-        let cleanup = bootstrap_pid_fallback(pidfd, pid, deadline, role);
-        close_fd(pidfd);
-        return cleanup.and(Err(RuntimeFingerprintProduceError::ContainmentUnavailable(
-            ContainmentUnavailableReason::PidfdUnavailable,
-        )));
-    }
-    match waitid_pidfd(pidfd, false, deadline) {
-        Ok((seen, libc::CLD_EXITED, 0)) if seen == pid => {
-            close_fd(pidfd);
-            Ok(())
-        }
-        Ok(_) => {
-            close_fd(pidfd);
-            Err(RuntimeFingerprintProduceError::ContainmentUnavailable(
-                ContainmentUnavailableReason::PidfdUnavailable,
-            ))
-        }
-        Err(()) => {
-            let cleanup = bootstrap_pid_fallback(pidfd, pid, deadline, role);
-            close_fd(pidfd);
-            cleanup.and(Err(RuntimeFingerprintProduceError::ContainmentUnavailable(
-                ContainmentUnavailableReason::PidfdUnavailable,
-            )))
-        }
-    }
-}
-
-fn bootstrap_pid_fallback(
-    pidfd: libc::c_int,
-    pid: libc::pid_t,
-    deadline: Instant,
-    role: super::RuntimeOwnedChildRole,
-) -> Result<(), RuntimeFingerprintProduceError> {
-    // The capability child is the sole registered-child bootstrap exception. Keep pidfd held.
-    if pidfd < 0 {
-        return Err(RuntimeFingerprintProduceError::ContainmentUnavailable(
-            ContainmentUnavailableReason::PidfdUnavailable,
-        ));
-    }
-    // SAFETY: pid is the exact capability direct child and cannot be reused before reap.
-    let signal_result = unsafe { libc::kill(pid, libc::SIGKILL) };
-    if signal_result != 0 && last_errno() != libc::ESRCH {
-        return Err(
-            RuntimeFingerprintProduceError::ChildRegistrationCleanupIncomplete {
-                role,
-                operation: super::RuntimeChildCleanupOperation::Termination,
-            },
-        );
-    }
-    let mut status = 0;
-    loop {
-        // SAFETY: pid is the capability direct child and pidfd remains held until return.
-        let result = unsafe { libc::waitpid(pid, &mut status, libc::WNOHANG) };
-        if result == pid {
-            return Ok(());
-        }
-        if result < 0 || Instant::now() >= deadline {
-            return Err(
-                RuntimeFingerprintProduceError::ChildRegistrationCleanupIncomplete {
-                    role,
-                    operation: super::RuntimeChildCleanupOperation::Reap,
-                },
-            );
-        }
-        std::thread::yield_now();
-    }
-}
-
 pub(super) fn waitid_pidfd(
     pidfd: libc::c_int,
     nowait: bool,
@@ -510,47 +208,6 @@ pub(super) fn waitid_pidfd(
         }
         std::thread::yield_now();
     }
-}
-
-fn child_capability_main(gate: [libc::c_int; 2], status: [libc::c_int; 2]) -> ! {
-    if !child_reset_signal_dispositions() {
-        child_write_status_and_exit(status[1], CHILD_SIGNAL_FAILED);
-    }
-    let descriptor_status = child_isolate_descriptors(gate[0], status[1]);
-    if descriptor_status != CHILD_READY {
-        child_write_status_and_exit(status[1], descriptor_status);
-    }
-    // SAFETY: these two inherited foreign ends are outside the allowlist and must be closed.
-    let foreign_gate = unsafe { libc::fcntl(gate[1], libc::F_GETFD) };
-    let foreign_status = unsafe { libc::fcntl(status[0], libc::F_GETFD) };
-    if foreign_gate != -1 || foreign_status != -1 || last_errno() != libc::EBADF {
-        child_write_status_and_exit(status[1], CHILD_DESCRIPTOR_FAILED);
-    }
-    let empty: u64 = 0;
-    // SAFETY: Linux x86_64/aarch64 use one u64 kernel signal set for rt_sigprocmask.
-    let mask_result = unsafe {
-        libc::syscall(
-            libc::SYS_rt_sigprocmask,
-            libc::SIG_SETMASK,
-            &empty,
-            std::ptr::null_mut::<u64>(),
-            std::mem::size_of::<u64>(),
-        )
-    };
-    if mask_result != 0 {
-        child_write_status_and_exit(status[1], CHILD_SIGNAL_FAILED);
-    }
-    if write_byte(status[1], CHILD_READY).is_err() {
-        // SAFETY: no Rust destructors may run in the post-fork child.
-        unsafe { libc::_exit(112) };
-    }
-    let mut go = 0_u8;
-    // SAFETY: gate[0] is the sole live gate read descriptor and points to one writable byte.
-    let read = unsafe { libc::read(gate[0], (&mut go as *mut u8).cast(), 1) };
-    close_fd(gate[0]);
-    close_fd(status[1]);
-    // SAFETY: the capability child performs no workload beyond the validated gate.
-    unsafe { libc::_exit(if read == 1 && go == CHILD_GO { 0 } else { 113 }) }
 }
 
 pub(super) fn child_reset_signal_dispositions() -> bool {
@@ -588,41 +245,6 @@ pub(super) fn child_reset_signal_dispositions() -> bool {
         signal += 1;
     }
     true
-}
-
-fn child_isolate_descriptors(first: libc::c_int, second: libc::c_int) -> u8 {
-    let low = first.min(second) as u32;
-    let high = first.max(second) as u32;
-    let ranges = [
-        (0_u32, low.saturating_sub(1), low != 0),
-        (low + 1, high.saturating_sub(1), high > low + 1),
-        (high + 1, u32::MAX, high != u32::MAX),
-    ];
-    for (start, end, required) in ranges {
-        if !required {
-            continue;
-        }
-        // SAFETY: close_range affects only the fork child and excludes both allowlisted fds.
-        if unsafe { libc::syscall(libc::SYS_close_range, start, end, 0) } != 0 {
-            let errno = last_errno();
-            return if matches!(
-                errno,
-                libc::ENOSYS | libc::EPERM | libc::EACCES | libc::EINVAL
-            ) {
-                CHILD_DESCRIPTOR_UNAVAILABLE
-            } else {
-                CHILD_DESCRIPTOR_FAILED
-            };
-        }
-    }
-    CHILD_READY
-}
-
-fn child_write_status_and_exit(fd: libc::c_int, status: u8) -> ! {
-    let _result = write_byte(fd, status);
-    close_fd(fd);
-    // SAFETY: no Rust destructors may run in the post-fork child.
-    unsafe { libc::_exit(111) }
 }
 
 pub(super) fn read_byte_before(
